backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
beanie>=1.26.0
motor>=3.4.0
pydantic>=2.10
pydantic[email]
pydantic-settings>=2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
aiofiles>=23.2.1
httpx>=0.27.0
